moved utility methods from dictIO/dict.py to dictIO/utils/dict.py
